Extract "memberof" tests into separate test class
