+++ drools-compiler/src/test/resources/org/drools/compiler/integrationtests/operators/memberof/test_memberOf.drl
